Configurable PVC name
Before the name of the default PVC for data was set to "data". This
could conflict when users add additional volumes and/or mounts that use
the same name, resulting in "duplicate volume" errors that are not as
easy to debug and understand in the first place.

To prevent such situations, the name of the PVC has been added as value
field keeping "data" as default name, but now allows users to change it
when required.