fix(pvc): prevents helm from removing a PVC
arguably, setting the reclaim policy on either StorageClasses or
PersistentVolumes would be the best way to ensure removal of a PVC
doesn't lead to accidental data loss. Though this can not be
configured in any object this Chart would install.

As an alternative, we may prevent Helm from ever deleting a
PersistentVolumeClaim, adding some `helm.sh/resource-policy=keep`
annotation.