CRD: remove 'status:' key from generated configs
The CRDs in this Helm chart all follow the form:

    kind: CustomResourceDefinition
    metadata:
      # ...
    spec:
      # ...
    status:
      acceptedNames:
        kind: ""
        plural: ""
      conditions: []
      storedVersions: []

The problem with the `status:` key is that it is usually
runtime-populated, for example by a controller.

In the case of a cluster managed by Anthos Config Management, it fails
to apply the Helm-generated Traefik configs with the error:

    KNV1045: Configs with "status" specified are not allowed.
    To fix, either remove the config or remove the "status" field in the config.
    For more information, see https://g.co/cloud/acm-errors#knv1045

If you remove the `status:` key from the CRDs, this now applies cleanly to an
Anthos-managed cluster.