Fix test for #669
`spec.template.spec.containers[0].args` is an array, so `notContains` is
testing that the given string does not exist in the array, not whether
it is present as an arbitrary substring.

Issues like this can be detected before release by following a
"red-green" approach when writing tests (e.g. write a failing test, like
this one, then make the code change to fix it).