Update DigitalOcean PROXY protocol documentation
When externalTrafficPolicy is set to Cluster, inbound requests from the load balancer could go to any node in the cluster. If they arrive at a node not running Traefik, they are port forwarded on to a node that is. That port forwarding has the effect of changing the source IP address of the inbound packets. In this circumstance, the PROXY protocol headers are rejected unless the cluster nodes are in the DOTrustedIPs.